requirements.txt
# Core dependencies
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2

# LLM and langchain related - fixed versions for compatibility
langchain
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langgraph
google-generativeai

# For vector database and embeddings
faiss-cpu==1.7.4
# sentence-transformers==2.2.0
# huggingface-hub==0.20.3
# torch

# Audio processing
openai-whisper==20231117
faster-whisper

# Utilities
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6